backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-session>=0.5.0
flask-sqlalchemy>=3.1.0
langchain>=1.2.10
langchain-community>=0.4.1
langchain-core>=1.2.11
mysql-connector-python>=9.6.0
google-generativeai>=0.8.4
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=2.2.4
cryptography>=41.0.0
